requirements.txt
git+https://github.com/yupix/Mi.py.git
loguru
python-dotenv
PyYAML
SQLAlchemy
alembic
git+https://github.com/yupix/dbmanager@93258484bcfe90a6ca10850cb121cfb993ec1039
